+++ src/categories/categories.controller.ts
@@ -6,6 +6,7 @@ import {
   Patch,
   Param,
   Delete,
+  NotFoundException,
 } from '@nestjs/common';
 import { CategoriesService } from './categories.service';
 import { CreateCategoryDto } from './dto/create-category.dto';
@@ -32,22 +33,26 @@ export class CategoriesController {
 
   @Get(':id')
   @ApiOkResponse({ type: CategoryEntity })
-  findOne(@Param('id') id: string) {
-    return this.categoriesService.findOne(id);
+  async findOne(@Param('id') id: string) {
+    const category = this.categoriesService.findOne(id);
+    if (!category) throw new NotFoundException(`Category ${id} not found`);
+    return category;
   }
 
   @Patch(':id')
   @ApiOkResponse({ type: CategoryEntity })
-  update(
+  async update(
     @Param('id') id: string,
     @Body() updateCategoryDto: UpdateCategoryDto,
   ) {
-    return this.categoriesService.update(id, updateCategoryDto);
+    const exCategory = await this.findOne(id);
+    if (exCategory) return this.categoriesService.update(id, updateCategoryDto);
   }
 
   @Delete(':id')
   @ApiOkResponse({ type: CategoryEntity })
-  remove(@Param('id') id: string) {
-    return this.categoriesService.remove(id);
+  async remove(@Param('id') id: string) {
+    const exCategory = await this.findOne(id);
+    if (exCategory) return this.categoriesService.remove(id);
   }
 }
