+++ src/users/users.controller.ts
@@ -60,7 +60,9 @@ export class UsersController {
   @ApiBearerAuth()
   @ApiOkResponse({ type: UserEntity })
   async update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
-    return new UserEntity(await this.usersService.update(id, updateUserDto));
+    const user = await this.findOne(id);
+    if (user)
+      return new UserEntity(await this.usersService.update(id, updateUserDto));
   }
 
   @Delete(':id')
@@ -68,6 +70,7 @@ export class UsersController {
   @ApiBearerAuth()
   @ApiOkResponse({ type: UserEntity })
   async remove(@Param('id') id: string) {
-    return new UserEntity(await this.usersService.remove(id));
+    const user = await this.findOne(id);
+    if (user) return new UserEntity(await this.usersService.remove(id));
   }
 }
